+++ src/pages/user/User.scss
@@ -9,121 +9,9 @@
 
     background: #111;
 
-    .main {
-        display: grid;
-        place-items: center;
-        grid-template-columns: auto 1fr;
+    // HeaderInfo
 
-        gap: 1rem;
-        padding-top: 0.75rem;
-        padding-left: 1rem;
+    // ConectionButtons
 
-        & img {
-            height: 4rem;
-            border: 1px solid #222;
-            border-radius: 50%;
-        }
-        .stats {
-            display: flex;
-            align-items: center;
-            justify-content: space-evenly;
-
-            gap: 1rem;
-
-            width: 100%;
-
-            & section {
-                display: grid;
-                place-items: center;
-
-                & h3 {
-                    font-size: 1rem;
-                    font-weight: 700;
-                }
-                & h4 {
-                    font-size: 0.75rem;
-                    font-weight: 500;
-                    color: #aaa;
-                }
-            }
-        }
-    }
-    .info {
-        display: grid;
-        gap: 0.5rem;
-        padding: 1rem;
-
-        width: 90%;
-
-        & h2 {
-            font-size: 1rem;
-            font-weight: 700;
-        }
-        & p {
-            font-size: 0.9rem;
-            font-weight: 400;
-        }
-    }
-    .actionButtons {
-        display: flex;
-        flex-wrap: wrap;
-        align-items: center;
-        justify-content: center;
-
-        gap: 0.25rem;
-        padding: 0 0.25rem;
-
-        & button,
-        a {
-            flex: 1;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-
-            gap: 0.25rem;
-            padding: 0.5rem;
-
-            font-size: 1rem;
-
-            width: 100%;
-            min-width: 200px;
-
-            border-radius: 2.5px;
-
-            background: #222;
-
-            & h4 {
-                text-transform: capitalize;
-                font-weight: 600;
-            }
-        }
-
-        & a {
-            transition: background-color 150ms ease-in-out;
-        }
-        a > svg,
-        a > h4 {
-            transition: color 150ms ease-in-out;
-        }
-
-        .followButton {
-            min-width: 275px;
-        }
-        .unfollowed {
-            background-color: var(--main);
-        }
-
-        .followed,
-        .instagram,
-        .twitter {
-            flex: 1;
-
-            color: #f5f5f5;
-            background-color: #222;
-
-            &:hover {
-                background-color: #111;
-            }
-        }
-    }
+    // Gallery
 }
