+++ src/pages/user/components/ConectionButtons.scss
@@ -1,0 +1,62 @@
+.ConectionButtons {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+    justify-content: center;
+
+    gap: 0.25rem;
+    padding: 0 0.25rem;
+
+    & button,
+    a {
+        flex: 1;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+
+        gap: 0.25rem;
+        padding: 0.5rem;
+
+        font-size: 1rem;
+
+        width: 100%;
+        min-width: 200px;
+
+        border-radius: 2.5px;
+
+        background: #222;
+
+        & h4 {
+            text-transform: capitalize;
+            font-weight: 600;
+        }
+    }
+
+    & a {
+        transition: background-color 150ms ease-in-out;
+    }
+    a > svg,
+    a > h4 {
+        transition: color 150ms ease-in-out;
+    }
+
+    .followButton {
+        min-width: 275px;
+    }
+    .unfollowed {
+        background-color: var(--main);
+    }
+
+    .followed,
+    .instagram,
+    .twitter {
+        flex: 1;
+
+        color: #f5f5f5;
+        background-color: #222;
+
+        &:hover {
+            background-color: #111;
+        }
+    }
+}
