+++ src/pages/user/components/HeaderInfo.scss
@@ -1,0 +1,79 @@
+.HeaderInfo {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+
+    gap: 1rem;
+    padding: 1rem;
+
+    & img {
+        flex: 1;
+        max-width: 125px;
+        width: 100%;
+
+        border: 1px solid #222;
+        border-radius: 50%;
+    }
+
+    .main {
+        flex: 3;
+        min-width: 300px;
+
+        display: grid;
+        gap: 0.5rem;
+
+        .stats {
+            display: flex;
+            align-items: center;
+
+            gap: 1.5rem;
+
+            & section {
+                display: flex;
+                flex-wrap: wrap;
+                justify-content: center;
+                align-items: center;
+
+                gap: 0.25rem;
+
+                & h3 {
+                    font-size: 1rem;
+                    font-weight: 700;
+                }
+                & h4 {
+                    font-size: 0.75rem;
+                    font-weight: 500;
+                    color: #888;
+                }
+            }
+        }
+    }
+    .info {
+        display: grid;
+        gap: 0.5rem;
+
+        & h2 {
+            display: flex;
+            align-items: center;
+
+            gap: 0.25rem;
+
+            .name {
+                text-transform: uppercase;
+                font-size: 1rem;
+                font-weight: 700;
+            }
+            .username {
+                font-size: 0.75rem;
+                font-weight: 400;
+                color: #888;
+            }
+        }
+        & p {
+            width: 90%;
+
+            font-size: 0.9rem;
+            font-weight: 400;
+        }
+    }
+}
